+++ history_tutorial_wsl.sh
@@ -128,13 +128,12 @@ show_communication_info() {
 # Download the file if needed
 download_and_open() {
   echo -e "\n${g}====================Trying to download the latest communication file====================${t}"
-  if curl -O -L "$github_url"; then
+  if curl -s --head --max-time 5 https://google.com > /dev/null && curl --max-time 10 -O -L "$github_url"; then
     cp "$communication_file" "$backup_file" 2>/dev/null
     show_communication_info
     explorer.exe "$communication_file" > /dev/null 2>&1 &
-  else
-    echo -e "\n\n${m}To receive current communication, please check your internet connection and try again next time!${t}" >&2
   fi
+  clear
 }
 
 # File paths
